backend/requirements.txt
alembic==1.9.2
anyio==3.6.2
asyncpg==0.27.0
bcrypt==4.0.1
black==23.1.0
certifi==2022.12.7
cffi==1.15.1
click==8.1.3
cryptography==3.4.8
dnspython==2.3.0
ecdsa==0.18.0
email-validator==1.3.1
fastapi==0.89.1
flake8==5.0.4
flake8-black==0.3.6
greenlet==2.0.2
h11==0.14.0
httpcore==0.16.3
httpx==0.23.3
idna==3.4
install==1.3.5
isort==5.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.4
MarkupSafe==2.1.2
mccabe==0.7.0
mypy-extensions==1.0.0
packaging==23.0
passlib==1.7.4
pathspec==0.11.0
platformdirs==3.0.0
psycopg2-binary==2.9.5
pyasn1==0.4.8
pycodestyle==2.9.1
pycparser==2.21
pydantic==1.10.4
pydot==1.4.2
pyflakes==2.5.0
PyJWT==1.7.1
pyparsing==3.0.9
python-jose==3.3.0
python-multipart==0.0.5
PyYAML==6.0
rfc3986==1.5.0
rsa==4.9
six==1.16.0
sniffio==1.3.0
sqladmin==0.8.0
SQLAlchemy==1.4.46
starlette==0.22.0
tomli==2.0.1
typing_extensions==4.4.0
uvicorn==0.20.0
WTForms==3.0.1
